requirements.txt
absl-py
accelerate
aiofiles
aiohappyeyeballs
aiohttp
aioredis
aiosignal
albucore
albumentations
alembic
annotated-types
antlr4-python3-runtime
anyio
argon2-cffi
argon2-cffi-bindings
arxiv
ascii_colors
async-timeout
asyncpg
attrs
audioread
backoff
bcrypt
beautifulsoup4
blinker
blis
boto3
botocore
Brotli
build
CacheControl
cachetools
catalogue
certifi
cffi
charset-normalizer
chromadb
click
cloudpathlib
coloredlogs
colorlog
confection
config
configparser
contourpy
cryptography
cssselect
cycler
cymem
dataclasses-json
decorator
Deprecated
deprecation
dill
diskcache
distro
dnspython
doclayout_yolo
dotenv
duckduckgo_search
durationpy
ecdsa
effdet
email-validator
emoji
et_xmlfile
faiss-cpu
fast-langdetect
fastapi
fasttext-predict
feedfinder2
feedparser
ffmpy
filelock
filetype
firebase_admin
Flask
flask-cors
flatbuffers
fonttools
frozenlist
fsspec
ftfy
future
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-cloud-vision
google-crc32c
google-generativeai
google-resumable-media
googleapis-common-protos
googlesearch-python
GPUtil
gradio
gradio_client
gradio_pdf
groovy
grpcio
grpcio-status
h11
h2
hf-xet
hpack
html5lib
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
hyperframe
idna
imageio
importlib_metadata
importlib_resources
itsdangerous
jieba3k
Jinja2
jiter
jmespath
joblib
json_repair
jsonlines
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kiwisolver
kubernetes
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langcodes
langdetect
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
language_data
lazy_loader
librosa
lightrag
lightrag-hku
llvmlite
logger
loguru
lxml
magika
Mako
marisa-trie
Markdown
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
mineru
mineru_vl_utils
ml_dtypes
mmh3
modelscope
mpmath
msgpack
msoffcrypto-tool
multidict
murmurhash
mypy_extensions
nano-vectordb
narwhals
nest-asyncio
networkx
newspaper3k
nltk
numba
numpy
nvidia-ml-py
oauthlib
olefile
omegaconf
onnx
onnxruntime
openai
opencv-python
opencv-python-headless
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
ormsgpack
overrides
packaging
pandas
passlib
pdf2image
pdfminer.six
pdfplumber
pdftext
pgvector
pi_heif
pikepdf
pillow
pip
pipmaster
platformdirs
plotly
polars
polars-runtime-32
pooch
postgrest
posthog
preshed
primp
prometheus_client
prometheus-fastapi-instrumentator
propcache
proto-plus
protobuf
psutil
psycopg2
psycopg2-binary
py-cpuinfo
pyasn1
pyasn1_modules
pybase64
pyclipper
pycocotools
pycparser
pydantic
pydantic_core
pydantic-settings
pydub
Pygments
PyJWT
PyMuPDF
pynvml
pypandoc
pyparsing
pypdf
PyPDF2
pypdfium2
pyphen
PyPika
pypinyin
pyproject_hooks
python-dateutil
python-docx
python-dotenv
python-iso639
python-jose
python-magic
python-multipart
python-oxmsg
python-pptx
pytz
PyYAML
raganything
rank-bm25
RapidFuzz
realtime
redis
referencing
regex
reportlab
requests
requests-file
requests-oauthlib
requests-toolbelt
rich
robust-downloader
rpds-py
rsa
ruff
s3transfer
safehttpx
safetensors
scikit-image
scikit-learn
scipy
seaborn
semantic-version
sentence-transformers
serpapi
setuptools
sgmllib3k
shapely
shellingham
simsimd
six
smart_open
sniffio
soundfile
soupsieve
soxr
spaces
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
srsly
sse-starlette
stanza
starlette
storage3
StrEnum
stringzilla
supabase
supabase-auth
supabase-functions
sympy
tenacity
tensorboard
tensorboard-data-server
textstat
thinc
thop
threadpoolctl
tifffile
tiktoken
timm
tinysegmenter
tldextract
tokenizers
tomlkit
torch
torchvision
tqdm
transformers
typer
typing_extensions
typing-inspect
typing-inspection
tzdata
ultralytics
ultralytics-thop
unstructured
unstructured-client
unstructured-inference
unstructured.pytesseract
uritemplate
urllib3
uvicorn
uvloop
wasabi
watchfiles
wcwidth
weasel
webencodings
websocket-client
websockets
Werkzeug
wikipedia
wrapt
xlrd
xlsxwriter
xxhash
yarl
zipp
zstandard
# Whisper transcription (Hugging Face Transformers) - ver chatbot/requirements.txt
# transformers>=4.40.0
# datasets[audio]>=2.14.0
# accelerate>=0.20.0
# torch>=2.0.0
# torchaudio>=2.0.0

